Scripts: New script to preprocess recorded databases
With this script we can do some preprocessing on the recorded databases.
At the moment it just duplicates the first channel of a WAV file into
the second one. Making a mono file into a stereo file with same content
on both channels.

Also the progress bar is pretty neat, kinda proud of it :^)