Scripts: changed sampling frequency and channels, also added metadata
I've changed the sampling frequency to 96 kHz and the number of channels
to 1 (mono). We are recording the speech with this high frequency in
case later we need to down sample to 16000 or 22050 Hz. Since we don't
know yet which one to use.

Also added the metadata file we are using in conjunction with the
script.