Scripts: Small python script to help us with the train queues
The way it works it's actually pretty simple, there is a file where each
line is a command to run. This script reads that file, if there is a
command to run: NICE, just run it. If there is no command just wait X
amount of time and try again!

Another script I'm quite proud of :^)